refactor: co-locate host-service hooks in renderer/hooks/host-service
Move all host-service-related hooks to renderer/hooks/host-service/
following repo structure guidelines (one hook per file, folder per hook,
matching names). These will eventually move to a shared package.

- Split useWorkspaceEvent.ts into useWorkspaceEvent/ and useWorkspaceHostUrl/
- Rename useDashboardDiffStats/ → useDiffStats/ (match hook name)
- Move useFileTree/ and useFileDocument/ from v2-workspace route
- Update all consumer import paths
