+++ apps/desktop/src/renderer/hooks/host-service/useDiffStats/index.ts
@@ -1,0 +1,1 @@
+export { type DiffStats, useDiffStats } from "./useDiffStats";
