+++ apps/desktop/src/renderer/hooks/host-service/useWorkspaceEvent/index.ts
@@ -1,0 +1,1 @@
+export { useWorkspaceEvent } from "./useWorkspaceEvent";
