+++ apps/desktop/src/renderer/hooks/host-service/useWorkspaceEvent/useWorkspaceEvent.ts
@@ -1,12 +1,8 @@
 import { getEventBus } from "@superset/workspace-client";
 import type { FsWatchEvent } from "@superset/workspace-fs/client";
-import { eq } from "@tanstack/db";
-import { useLiveQuery } from "@tanstack/react-db";
-import { useEffect, useEffectEvent, useMemo } from "react";
+import { useEffect, useEffectEvent } from "react";
 import { getHostServiceWsToken } from "renderer/lib/host-service-auth";
-import { getRemoteHostUrl } from "renderer/lib/v2-workspace-host";
-import { useCollections } from "renderer/routes/_authenticated/providers/CollectionsProvider";
-import { useHostService } from "renderer/routes/_authenticated/providers/HostServiceProvider";
+import { useWorkspaceHostUrl } from "../useWorkspaceHostUrl";
 
 /**
  * Subscribe to an event bus event for a workspace.
@@ -66,39 +62,4 @@ export function useWorkspaceEvent(
 			}
 		};
 	}, [enabled, hostUrl, type, workspaceId]);
-}
-
-/**
- * Resolves a workspace ID to its host-service URL.
- * Local host → localhost port. Remote host → relay proxy URL.
- */
-export function useWorkspaceHostUrl(workspaceId: string): string | null {
-	const collections = useCollections();
-	const { services } = useHostService();
-
-	const { data: workspaceWithHost = [] } = useLiveQuery(
-		(q) =>
-			q
-				.from({ workspaces: collections.v2Workspaces })
-				.innerJoin({ hosts: collections.v2Hosts }, ({ workspaces, hosts }) =>
-					eq(workspaces.hostId, hosts.id),
-				)
-				.where(({ workspaces }) => eq(workspaces.id, workspaceId))
-				.select(({ workspaces, hosts }) => ({
-					hostId: workspaces.hostId,
-					hostOrgId: hosts.organizationId,
-				})),
-		[collections, workspaceId],
-	);
-
-	const match = workspaceWithHost[0] ?? null;
-
-	return useMemo(() => {
-		if (!match) return null;
-		// If we have a local host-service for this org, use it directly
-		const localService = services.get(match.hostOrgId);
-		if (localService) return localService.url;
-		// Otherwise route through the relay proxy
-		return getRemoteHostUrl(match.hostId);
-	}, [match, services]);
 }
