+++ apps/desktop/src/renderer/hooks/host-service/useWorkspaceHostUrl/index.ts
@@ -1,0 +1,1 @@
+export { useWorkspaceHostUrl } from "./useWorkspaceHostUrl";
