--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/DashboardSidebar/hooks/useDashboardDiffStats/index.ts
@@ -1,1 +1,0 @@
-export { type DiffStats, useDiffStats } from "./useDashboardDiffStats";
