--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/DashboardSidebar/hooks/useWorkspaceEvent/index.ts
@@ -1,1 +1,0 @@
-export { useWorkspaceEvent, useWorkspaceHostUrl } from "./useWorkspaceEvent";
