+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/FilePane/FilePane.tsx
@@ -1,8 +1,8 @@
 import type { RendererContext } from "@superset/panes";
 import { useCallback } from "react";
+import { useFileDocument } from "renderer/hooks/host-service/useFileDocument";
 import { isImageFile, isMarkdownFile } from "shared/file-types";
 import type { FilePaneData, PaneViewerData } from "../../../../types";
-import { useFileDocument } from "../../../useFileDocument";
 import { CodeRenderer } from "./renderers/CodeRenderer";
 import { ImageRenderer } from "./renderers/ImageRenderer";
 import { MarkdownRenderer } from "./renderers/MarkdownRenderer";
